requirements.txt
# =====================================================
# Production Dependencies for Student Platform API
# Python 3.11+
# =====================================================

# -----------------------------------------------------
# Core Flask Framework
# -----------------------------------------------------
Flask==3.1.2
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.2.1
blinker==1.9.0

# -----------------------------------------------------
# Database & ORM
# -----------------------------------------------------
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
Flask-Migrate==4.0.5
Mako==1.3.10

# -----------------------------------------------------
# API & Serialization
# -----------------------------------------------------
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
flask-marshmallow==0.15.0
flask-restx==1.3.0
aniso8601==10.0.1

# -----------------------------------------------------
# API Documentation
# -----------------------------------------------------
flasgger==0.9.7.1
apispec==6.3.1
apispec-webframeworks==0.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
referencing==0.36.2
rpds-py==0.27.1

# -----------------------------------------------------
# Security & Authentication
# -----------------------------------------------------
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
python-jose==3.3.0
cryptography==41.0.7
limits==5.5.0

# -----------------------------------------------------
# Configuration & Environment
# -----------------------------------------------------
python-dotenv==1.0.1
python-decouple==3.8

# -----------------------------------------------------
# Utilities
# -----------------------------------------------------
python-dateutil==2.8.2
pytz==2025.2
tzdata==2025.2

# -----------------------------------------------------
# Validation
# -----------------------------------------------------
email-validator==2.1.0
phonenumbers==8.13.27

# -----------------------------------------------------
# Caching (Optional but recommended)
# -----------------------------------------------------
redis==5.0.1
Flask-Caching==2.1.0
cachelib==0.9.0

# -----------------------------------------------------
# Logging & Monitoring
# -----------------------------------------------------
python-json-logger==2.0.7
sentry-sdk[flask]==1.39.1

# -----------------------------------------------------
# Production Server
# -----------------------------------------------------
gunicorn==21.2.0
gevent==23.9.1
greenlet==3.2.4

# -----------------------------------------------------
# HTTP Client (for webhooks/external APIs)
# -----------------------------------------------------
requests==2.31.0
urllib3==2.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10

# -----------------------------------------------------
# Task Queue (Optional)
# -----------------------------------------------------
celery==5.3.4
kombu==5.3.4
billiard==4.2.1
vine==5.1.0
amqp==5.3.1

# -----------------------------------------------------
# Testing
# -----------------------------------------------------
pytest==7.4.3
pytest-flask==1.3.0
attrs==25.3.0
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0

# -----------------------------------------------------
# Development Tools (Optional)
# -----------------------------------------------------
black==25.1.0
mypy-extensions==1.1.0
pathspec==0.12.1
platformdirs==4.4.0
rich==13.9.4
typing-extensions==4.15.0
